examples/python-chariott/requirements.txt
grpcio==1.50.0
grpcio-tools==1.50.0
pkg_resources==0.0.0
protobuf==4.21.9
six==1.16.0
